Address @nilsding's review comments
- Fix types on legacy attachment migrations
- Restore Action Mailer setup for development env
- Use `Hash#dig` in CarrierWave initializer

+++ db/migrate/20141229085904_add_attachment_profile_picture_to_users.rb
@@ -3,7 +3,7 @@ class AddAttachmentProfilePictureToUsers < ActiveRecord::Migration[4.2]
     change_table :users do |t|
       t.string :profile_picture_file_name
       t.string :profile_picture_content_type
-      t.bigint :profile_picture_file_size
+      t.integer :profile_picture_file_size
       t.datetime :profile_picture_updated_at
     end
   end
